# Cache package lock when running tests to speed up installation