--- a/test/e2e/test-utils-tests/basic/basic.test.ts
+++ b/test/e2e/test-utils-tests/basic/basic.test.ts
@@ -1,19 +1,14 @@
-import { createNext } from 'e2e-utils'
-import { NextInstance } from 'test/lib/next-modes/base'
-import { fetchViaHTTP } from 'next-test-utils'
+import { createNextDescribe } from 'e2e-utils'
 
-describe('createNext', () => {
-  let next: NextInstance
-
-  beforeAll(async () => {
-    next = await createNext({
-      files: __dirname,
+createNextDescribe(
+  'createNextDescribe',
+  {
+    files: __dirname,
+  },
+  ({ next }) => {
+    it('should work', async () => {
+      const res = await next.fetch('/')
+      expect(await res.text()).toContain('Hello World')
     })
-  })
-  afterAll(() => next.destroy())
-
-  it('should work', async () => {
-    const res = await fetchViaHTTP(next.url, '/')
-    expect(await res.text()).toContain('Hello World')
-  })
-})
+  }
+)