--- a/test/lib/create-next-install.js
+++ b/test/lib/create-next-install.js
@@ -212,7 +212,6 @@ async function createNextInstall({
             }
           })
       }
-      await new Promise(() => {})
 
       await fs.remove(tmpRepoDir)
       return installDir