- Cache `pnpm-lock.yaml` in `test/tmp`. It will do so just for tests that don't specify their own dependencies. Default dependencies work with typescript now so it shouldn't be an issue.
- Improve log messages a bit.

It reduces the time for installation on subsequent runs from `3.5s` to `.5s`
So we get this in dev mode:
```
createNext 2.3s
├─ init next dev instance 50 µs
├─ createTestDir 2s
│  ├─ createNextInstall 2s
│  │  ├─  enruse swc binary 4.4 ms
│  │  ├─ copy package.json to temp dir 1.1 ms
│  │  ├─ copy packages to temp dir 489 ms
│  │  ├─ linkPackages 762 ms
│  │  │  ├─ prepare packages for packing 38 ms
│  │  │  └─ packing packages 724 ms
│  │  │     ├─ pack create-next-app 587 ms
│  │  │     ├─ pack eslint-config-next 612 ms
│  │  │     ├─ pack @next/eslint-plugin-next 563 ms
│  │  │     ├─ pack @next/font 629 ms
│  │  │     ├─ pack next 563 ms
│  │  │     ├─ pack @next/bundle-analyzer 617 ms
│  │  │     ├─ pack @next/codemod 623 ms
│  │  │     ├─ pack @next/env 624 ms
│  │  │     ├─ pack @next/mdx 611 ms
│  │  │     ├─ pack @next/plugin-storybook 623 ms
│  │  │     ├─ pack @next/polyfill-module 614 ms
│  │  │     ├─ pack @next/polyfill-nomodule 602 ms
│  │  │     ├─ pack @next/swc 625 ms
│  │  │     ├─ pack @next/react-dev-overlay 586 ms
│  │  │     └─ pack @next/react-refresh-utils 628 ms
│  │  └─ run generic install command 665 ms
│  └─ writeInitialFiles 813 µs
└─ start next instance 281 ms
```