@JanKaifer 2023-01-03
Hmm, the error doesn't sem related but it also fails on my machine.