@ijjk 2023-01-03
# Stats from current PR

<details>
<summary><strong>Default Build</strong> (Decrease detected ✓)</summary>

<br/>

<details>
<summary><strong>General</strong> Overall decrease ✓</summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| buildDuration | 36.7s | 37.1s | ⚠️ +444ms |
| buildDurationCached | 15.8s | 11s | -4.8s |
| nodeModulesSize | 95.1 MB | 95.1 MB | ✓ |
| nextStartRea..uration (ms) | 279ms | 275ms | -4ms |

</details>

<details>
<summary><strong>Client Bundles (main, webpack)</strong></summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| 152-HASH.js gzip | 64.5 kB | 64.5 kB | ✓ |
| 53.HASH.js gzip | 182 B | 182 B | ✓ |
| main-app-HASH.js gzip | 205 B | 205 B | ✓ |
| main-HASH.js gzip | 79.2 kB | 79.2 kB | ✓ |
| webpack-HASH.js gzip | 1.7 kB | 1.7 kB | ✓ |
| Overall change | 146 kB | 146 kB | ✓ |

</details>

<details>
<summary><strong>Legacy Client Bundles (polyfills)</strong></summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| polyfills-HASH.js gzip | 31 kB | 31 kB | ✓ |
| Overall change | 31 kB | 31 kB | ✓ |

</details>

<details>
<summary><strong>Client Pages</strong></summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| _app-HASH.js gzip | 192 B | 192 B | ✓ |
| _error-HASH.js gzip | 179 B | 179 B | ✓ |
| amp-HASH.js gzip | 484 B | 484 B | ✓ |
| css-HASH.js gzip | 804 B | 804 B | ✓ |
| dynamic-HASH.js gzip | 2.27 kB | 2.27 kB | ✓ |
| edge-ssr-HASH.js gzip | 255 B | 255 B | ✓ |
| head-HASH.js gzip | 827 B | 827 B | ✓ |
| hooks-HASH.js gzip | 848 B | 848 B | ✓ |
| image-HASH.js gzip | 4.3 kB | 4.3 kB | ✓ |
| index-HASH.js gzip | 252 B | 252 B | ✓ |
| link-HASH.js gzip | 2.69 kB | 2.69 kB | ✓ |
| routerDirect..HASH.js gzip | 782 B | 782 B | ✓ |
| script-HASH.js gzip | 857 B | 857 B | ✓ |
| withRouter-HASH.js gzip | 781 B | 781 B | ✓ |
| 85e02e95b279..7e3.css gzip | 107 B | 107 B | ✓ |
| Overall change | 15.6 kB | 15.6 kB | ✓ |

</details>

<details>
<summary><strong>Client Build Manifests</strong></summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| _buildManifest.js gzip | 482 B | 482 B | ✓ |
| Overall change | 482 B | 482 B | ✓ |

</details>

<details>
<summary><strong>Rendered Page Sizes</strong></summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| index.html gzip | 489 B | 489 B | ✓ |
| link.html gzip | 504 B | 504 B | ✓ |
| withRouter.html gzip | 485 B | 485 B | ✓ |
| Overall change | 1.48 kB | 1.48 kB | ✓ |

</details>

<details>
<summary><strong>Edge SSR bundle Size</strong> Overall increase ⚠️</summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| edge-ssr.js gzip | 110 kB | 110 kB | ✓ |
| page.js gzip | 98.4 kB | 98.4 kB | ⚠️ +2 B |
| Overall change | 209 kB | 209 kB | ⚠️ +2 B |

</details>

<details>
<summary><strong>Middleware size</strong> Overall increase ⚠️</summary>

|  | vercel/next.js canary  | JanKaifer/next.js jankaifer/next-289-speed-up-installation-phase-of-running | Change |
| - | - | - | - |
| middleware-b..fest.js gzip | 581 B | 582 B | ⚠️ +1 B |
| middleware-r..fest.js gzip | 145 B | 145 B | ✓ |
| middleware.js gzip | 27 kB | 27 kB | ✓ |
| edge-runtime..pack.js gzip | 1.83 kB | 1.83 kB | ✓ |
| Overall change | 29.6 kB | 29.6 kB | ⚠️ +1 B |

</details>

#### Diffs
<details>
<summary>Diff for <strong>page.js</strong></summary>

```diff
@@ -8618,16 +8618,16 @@ Learn More: https://nextjs.org/docs/messages/node-module-in-edge-runtime`;
       /***/
     },
 
-    /***/ 1036: /***/ (
+    /***/ 199: /***/ (
       __unused_webpack_module,
       __unused_webpack_exports,
       __webpack_require__
     ) => {
       Promise.resolve(/* import() eager */).then(
-        __webpack_require__.bind(__webpack_require__, 1210)
+        __webpack_require__.bind(__webpack_require__, 2973)
       );
       Promise.resolve(/* import() eager */).then(
-        __webpack_require__.bind(__webpack_require__, 2973)
+        __webpack_require__.bind(__webpack_require__, 1210)
       );
       Promise.resolve(/* import() eager */).then(
         __webpack_require__.bind(__webpack_require__, 6448)
```
</details>
<details>
<summary>Diff for <strong>middleware-b..-manifest.js</strong></summary>

```diff
@@ -9,7 +9,7 @@ self.__BUILD_MANIFEST = {
   rootMainFiles: [
     "static/chunks/webpack-c452a3e31b73f504.js",
     "static/chunks/152-792c0b16c9234019.js",
-    "static/chunks/main-app-c9a0aaff35eafcb5.js"
+    "static/chunks/main-app-ec4ed0640c5bbc2a.js"
   ],
   pages: {
     "/": [
```
</details>
<details>
<summary>Diff for <strong>main-app-HASH.js</strong></summary>

```diff
@@ -1,16 +1,16 @@
 (self["webpackChunk_N_E"] = self["webpackChunk_N_E"] || []).push([
   [744],
   {
-    /***/ 9915: /***/ function(
+    /***/ 9421: /***/ function(
       __unused_webpack_module,
       __unused_webpack_exports,
       __webpack_require__
     ) {
       Promise.resolve(/* import() eager */).then(
-        __webpack_require__.t.bind(__webpack_require__, 429, 23)
+        __webpack_require__.t.bind(__webpack_require__, 1161, 23)
       );
       Promise.resolve(/* import() eager */).then(
-        __webpack_require__.t.bind(__webpack_require__, 1161, 23)
+        __webpack_require__.t.bind(__webpack_require__, 429, 23)
       );
       Promise.resolve(/* import() eager */).then(
         __webpack_require__.t.bind(__webpack_require__, 8138, 23)
@@ -28,7 +28,7 @@
       return __webpack_require__((__webpack_require__.s = moduleId));
     };
     /******/ __webpack_require__.O(0, [152], function() {
-      return __webpack_exec__(7070), __webpack_exec__(9915);
+      return __webpack_exec__(7070), __webpack_exec__(9421);
     });
     /******/ var __webpack_exports__ = __webpack_require__.O();
     /******/ _N_E = __webpack_exports__;
```
</details>
</details>
Commit: e33e1eccfa5c6f467f881053b0d5339ec0cce5cb